# ⚡️ Speed up function `_assign_hash_ids` by 34%