--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,6 +1,25 @@
+## 0.18.15-dev2
+
+### Enhancements
+- Speed up function _assign_hash_ids by 34% (codeflash)
+
+### Features
+
+### Fixes
+
+## 0.18.15-dev1
+
+### Enhancements
+- Speed up function group_broken_paragraphs by 30% (codeflash)
+
+### Features
+
+### Fixes
+
 ## 0.18.15-dev0
 
 ### Enhancements
+- Optimized the runtime of `ElementHtml._get_children_html`
 
 ### Features