--- a/unstructured/__version__.py
+++ b/unstructured/__version__.py
@@ -1,2 +1,1 @@
 __version__ = "0.18.15-dev1"  # pragma: no cover
-