--- a/unstructured/__version__.py
+++ b/unstructured/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.18.15"  # pragma: no cover
+__version__ = "0.18.16-dev0"  # pragma: no cover